% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCENTfunctions.R
\name{assoc_lmm}
\alias{assoc_lmm}
\title{Perform Mixed Linear regression: exprs ~ peak + (1|RandomEffect) + FixedEffect_covariates}
\usage{
assoc_lmm(data, idx = seq_len(nrow(data)), formula)
}
\arguments{
\item{data}{contains expr values and associated peak and covariates for a gene.}

\item{idx}{rows of the data to use: argument for boot function (bootstrapping)}

\item{formula}{user defined formula based on initialization in CreateSCENTObj Constructor}
}
\value{
vector: (coefficient of the peak effect on gene, variance of peak effect on gene)
}
\description{
Perform Mixed Linear regression: exprs ~ peak + (1|RandomEffect) + FixedEffect_covariates
}
