% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCENTfunctions.R
\name{basic_p}
\alias{basic_p}
\title{Derive a p-value from a vector of bootstrap samples using the "basic" calculation}
\usage{
basic_p(obs, boot, null = 0)
}
\arguments{
\item{obs}{observed value of parameter (using actual data)}

\item{boot}{vector of bootstraps}
}
\value{
p-value
}
\description{
Derive a p-value from a vector of bootstrap samples using the "basic" calculation
}
