% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCENTfunctions.R
\name{check_dimensions}
\alias{check_dimensions}
\title{Validity and Type Checking for CreateSCENTObject Constructor}
\usage{
check_dimensions(object)
}
\arguments{
\item{object}{SCENT object constructed from class CreateSCENTObject}
}
\value{
None OR Errors dependent on if the object follows the guidelines for SCENT
RNA: matrix of (genes x cells)
ATAC: matrix of (peaks x cells)
}
\description{
Validity and Type Checking for CreateSCENTObject Constructor
}
