% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{gene_hom}
\alias{gene_hom}
\title{Find the Homogeneity of a Gene Within a Population}
\usage{
gene_hom(expr, unit = "log2", normalise = TRUE, transpose = FALSE)
}
\arguments{
\item{expr}{A vector or matrix of gene expressions. For the matrix, genes
should be represented as rows and cells as columns.}

\item{unit}{The units to be parsed to the entropy function.}

\item{normalise}{A logical value representing whether the gene frequencies
should be normalised into a distribution.}

\item{transpose}{A legical value representing whether the matrix should be
transposed before any calculations are performed.}
}
\value{
A vector of the information contained in the distribution of each
gene. The higher this is, the more homogeneous the gene is within the cell
population.
}
\description{
Find the Homogeneity of a Gene Within a Population
}
