% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferDMAPandRoot.R
\name{InferDMAPandRoot}
\alias{InferDMAPandRoot}
\title{Infers Diffusion Map, Root-state and root-cell}
\usage{
InferDMAPandRoot(
  pot.v,
  exp.m,
  avth=1,
  sdth=0.25,
  kDMAP = 30,
  pctop = 0.05
)
}
\arguments{
\item{pot.v}{A vector of potency estimates for all cells, e.g. a vector of SR or CCAT values.}

\item{exp.m}{The normalized and log-transformed scRNA-Seq data matrix}

\item{avth}{The threshold on the average expression to use for selecting genes in the diffusion map. Default values is 1.}

\item{sdth}{The threshold on the standard deviation to use for selecting genes in the diffusion map. Default value is 0.25.}

\item{kDMAP}{An integer specifying the input argument to the diffusion map, representing the number of
 nearest neighbors to consider. By default this value is 30.}

\item{pctop}{The proportion of top-ranked cells (with cells ranked by SR or CCAT) to consider when inferring teh root-state. By default this values is 0.05.}
}
\value{
A list with the following elements:

\item{dmap}{The diffusion map object, as given by the `destiny` package.}

\item{dc}{The inferred diffusion map component matrix, with rows labeling cells and columns labeling ranked diffusion components.}

\item{transM}{The Markov transition matrix associated with the inferred diffusion map. Rows and columns label cells.}

\item{root}{The column index of the root-cell.}

\item{color}{A color vector for the cells stratified according to potency}
}
\description{
This function constructs the diffusion map, Markov transition matrix and uses the provided potency estimates to infer the root-state and root-cell.
}
\examples{


}
\references{
Teschendorff AE, Tariq Enver. 
\emph{Single-cell entropy for accurate estimation of differentiation 
potency from a cell’s transcriptome.}
Nature communications 8 (2017): 15599.
doi:\href{https://doi.org/10.1038/ncomms15599}{
10.1038/ncomms15599}.
}
