% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCENTfunctions.R
\name{interp_pval}
\alias{interp_pval}
\title{Interpolate a p-value from quantiles that should be "null scaled"}
\usage{
interp_pval(q)
}
\arguments{
\item{q}{bootstrap quantiles, centered so that under the null, theta = 0}
}
\value{
two-sided p-value
}
\description{
Interpolate a p-value from quantiles that should be "null scaled"
}
