% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{normalise}
\alias{normalise}
\title{Normalise Counts into a Distribution}
\usage{
normalise(dist)
}
\arguments{
\item{dist}{A vector of a frequency distribution.}
}
\value{
A vector of a probability distribution relative to the
frequencies.
}
\description{
A function that takes frequency count data and normalises it into a
probability distribution. Only available internally within SCEnt.
}
