% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCENT.R
\name{SCENT}
\alias{SCENT}
\title{SCENT}
\usage{
SCENT(SS, r, nn, rho = 1, Tol = 1e-04, Niter = 200)
}
\arguments{
\item{SS}{A data array of dimension p by p by K of covariance matrices}

\item{r}{A scalar of the prespecified rank of the tensor decomposition (an upper bound for # clusters)}

\item{nn}{A vector of length K of sample size for each population}

\item{rho}{A scalar of augmented parameter in ADMM, default=0.1}

\item{Tol}{Convergence tolerance for relative residuals, default=1E-4}

\item{Niter}{Maximum number of iterations, default=200}
}
\value{
primal, dual, obj, pen: each being a sequence of values corresponding to
     the primal residual, dual residual, objective value, and penalty value

Omega A data array of dimension p by p by K of precision matrices

Theta0 A matrix of size p by p, common network structure

Theta A tensor of size p by p by r (or a smaller rank, can be 0), each layer corresponds to a unit-norm, sparse sub-network

U A matrix of size K by r (or a smaller rank, can be 0), the nonsparsity of each column corresponds to the cluster index
}
\description{
This function uses the ADMM algorithm to conduct Simultaneous Clustering and Estimation of Networks
}
\examples{
SS=rWishart(6,20,0.01*diag(20))
out=SCENT(SS,r=2,nn=rep(100,6),
              rho=50)
par(mfrow=c(1,2))
plot(out$primal)
plot(out$dual)
out$U

}
\keyword{SCENT}
