% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{scent_select_tidy}
\alias{scent_select_tidy}
\title{A Tidy Wrapper for Feature Selection by Heterogeneity}
\usage{
scent_select_tidy(
  expr,
  bit_threshold = NULL,
  count_threshold = NULL,
  perc_threshold = NULL,
  unit = "log2",
  normalise = TRUE,
  transpose = FALSE
)
}
\arguments{
\item{expr}{A tibble of gene expression data. Cells should be represented
as rows and genes should be represented as columns.}

\item{bit_threshold}{The threshold for the amount of bits of information a
gene must add to be selected as a feature. Only one threshold can be used
at a time.}

\item{count_threshold}{A number represented how many of the most
heterogeneous cells should be selected. Only one threshold can be used at
a time.}

\item{perc_threshold}{The percentile of the hetergeneity distribution above
which a gene should be to be selected as a feature.}

\item{unit}{The units to be used when calculating entropy.}

\item{normalise}{A logical value representing whether the gene counts
should be normalised into a probability distribution.}

\item{transpose}{A logical value representing whether the matrix should
be transposed before having any operations computed on it.}
}
\value{
A tibble of gene expression values where genes with low
heterogeneity have been removed.
}
\description{
A Tidy Wrapper for Feature Selection by Heterogeneity
}
