% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCENTfunctions.R
\name{CreatePeakToGeneList}
\alias{CreatePeakToGeneList}
\title{Creating Cis Gene-Peak Pair Lists to Parallelize Through}
\usage{
CreatePeakToGeneList(
  object,
  genebed = "/path/to/GeneBody_500kb_margin.bed",
  nbatch,
  tmpfile = "./temporary_atac_peak.bed",
  intersectedfile = "./temporary_atac_peak_intersected.bed.gz"
)
}
\arguments{
\item{object}{SCENT object}

\item{genebed}{File directory for bed file that contains 500 kb windows for each gene}

\item{nbatch}{Number of batches to produce: Length of the list}

\item{tmpfile}{Location of temporary file.}

\item{intersectedfile}{Location of intersected file.}
}
\value{
SCENT object with updated field of peak.info
}
\description{
Creating Cis Gene-Peak Pair Lists to Parallelize Through
}
