% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{gene_het}
\alias{gene_het}
\title{Find the Heterogeneity of a Gene Within a Population}
\usage{
gene_het(expr, unit = "log2", normalise = TRUE, transpose = FALSE)
}
\arguments{
\item{expr}{A vector or matrix of gene expressions. For the matrix, genes
should be represented as rows and cells as columns.}

\item{unit}{The units to be parsed to the entropy function.}

\item{normalise}{A logical value representing whether the gene frequencies
should be normalised into a distribution.}

\item{transpose}{A logical value representing whether the matrix should be
transposed before any calculations are performed.}
}
\value{
A vector of the information gained from the gene distribution
compared to the uniform distribution. The higher the value more
heterogeneous the cell is within the population.
}
\description{
Find the Heterogeneity of a Gene Within a Population
}
