% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCENTfunctions.R
\docType{class}
\name{SCENT-class}
\alias{SCENT-class}
\alias{CreateSCENTObj}
\title{SCENT Class Constructor}
\value{
SCENT object to use for further analysis
}
\description{
SCENT Class Constructor
}
\section{Slots}{

\describe{
\item{\code{rna}}{dgCMatrix. scRNAseq matrix read as a sparse matrix}

\item{\code{atac}}{dgCMatrix. scATACseq matrix read as a sparse matrix}

\item{\code{meta.data}}{data.frame. Metadata table with covariates and a cell ID column ("cell")}

\item{\code{peak.info}}{data.frame. Dataframe that contains gene-peak pairs for SCENT to search through}

\item{\code{peak.info.list}}{list. List of dataframes that contain gene-peak pairs to parallelize through}

\item{\code{covariates}}{character. Assign covariates that are needed for the analysis. Must be names that are in the columns of meta.data}

\item{\code{random.effects}}{character. Assign covariates that are to be treated as random effects if using randome effects modeling. if not set to NULL}

\item{\code{celltypes}}{character. Assign celltype column from meta.data}

\item{\code{SCENT.result}}{data.frame. Initialized as empty. Becomes a table of resultant significant gene peak pairs}
}}

