% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCENTfunctions.R
\name{SCENT_algorithm}
\alias{SCENT_algorithm}
\title{SCENT Algorithm: Poisson Regression with Empirical P-values through Bootstrapping.}
\usage{
SCENT_algorithm(object, celltype, ncores)
}
\arguments{
\item{object}{SCENT object}

\item{celltype}{User specified cell type defined in celltypes column of meta.data}

\item{ncores}{Number of cores to use for Parallelization}
}
\value{
SCENT object with updated field SCENT.results
}
\description{
SCENT Algorithm: Poisson Regression with Empirical P-values through Bootstrapping.
}
